Changing to dark mode will stay in setting fragment now
1. Changing to dark mode will stay in the setting fragment now
2. Added test data to goalsTest.json
3. changed the addLog and addGoal to floatingActionButton

+++ app/src/main/java/com/example/csc306b_cw/GoalsData.kt
@@ -4,7 +4,7 @@ class GoalsData (
     var goalName: String? = null,
     var interval: Int? = null,
     var unit: String? = null,
-    var durationPerUnit: Int? = null,
+    var durationPerUnit: Double? = null,
     var progress: String? = null,
     var deadline: String? = null,
     var description: String? = null,
